# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: getProjects
      tags: [public]
      summary: List all projects
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [admin]
      summary: Create a project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [public]
      summary: Get single project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [admin]
      summary: Update project (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [admin]
      summary: Delete project (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Skills ────────────────────────────────────────────────────────────────
  /skills:
    get:
      operationId: getSkills
      tags: [public]
      summary: List skills
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [admin]
      summary: Create skill (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    patch:
      operationId: updateSkill
      tags: [admin]
      summary: Update skill (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [admin]
      summary: Delete skill (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Experience ────────────────────────────────────────────────────────────
...
  /testimonials:
    get:
      operationId: getTestimonials
      tags: [public]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [admin]
      summary: Create testimonial (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [admin]
      summary: Update testimonial (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [admin]
      summary: Delete testimonial (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Certificates ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, description, category, techStack, featured, order]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        category:
          type: string
        techStack:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: integer
        status:
          type: string
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [title, description, category, techStack]
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        category:
          type: string
        techStack:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        videoUrl:
          type: string
        liveUrl:
          type: string
        githubUrl:
          type: string
        featured:
          type: boolean
        order:
          type: integer
        status:
          type: string

    # ─── Skill ───────────────────────────────────────────────────────────────
    Skill:
      type: object
      required: [id, name, category, proficiency, icon, order]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        proficiency:
          type: integer
        icon:
          type: string
        color:
          type: ["string", "null"]
        order:
          type: integer
    SkillInput:
      type: object
      required: [name, category, proficiency, icon]
      properties:
        name:
          type: string
        category:
          type: string
        proficiency:
          type: integer
        icon:
          type: string
        color:
          type: string
        order:
          type: integer

    # ─── Experience ──────────────────────────────────────────────────────────
...
    Testimonial:
      type: object
      required: [id, name, role, company, content, rating]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        company:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        featured:
          type: boolean
    TestimonialInput:
      type: object
      required: [name, role, company, content, rating]
      properties:
        name:
          type: string
        role:
          type: string
        company:
          type: string
        content:
          type: string
        rating:
          type: integer
        avatarUrl:
          type: string
        linkedinUrl:
          type: string
        featured:
          type: boolean

    # ─── Certificate ─────────────────────────────────────────────────────────
...