# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  category: string;
  proficiency: number;
  icon: string;
  /** @nullable */
  color?: string | null;
  order: number;
}