# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  content: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  featured?: boolean;
}